Move message enums into appropriate files (no functional changes)
